Convert ConfigurationManager::[Get/Store]SetupDiscriminator to uint16_t

+++ examples/lock-app/nrf5/main/AppTask.cpp
@@ -173,7 +173,7 @@ int AppTask::Init()
         CHIP_ERROR err = CHIP_NO_ERROR;
         chip::SetupPayload payload;
         uint32_t setUpPINCode       = 0;
-        uint32_t setUpDiscriminator = 0;
+        uint16_t setUpDiscriminator = 0;
 
         err = ConfigurationMgr().GetSetupPinCode(setUpPINCode);
         if (err != CHIP_NO_ERROR)
